[refactor] 파일정리
