+++ src/Mobile.jsx
@@ -1,19 +1,19 @@
 import styled from "styled-components";
 import { Routes, Route } from "react-router-dom";
 import Home from "./pages/Home";
-import SignUp from "./pages/SignUp";
-import Login from "./pages/Login";
+import SignUp from "./pages/signup/SignUp";
+import Login from "./pages/signup/Login";
 import KakaoOauth from "./shared/kakaoOauth";
-import EmailSend from "./pages/EmailSend";
-import PwSend from "./pages/PwSend";
-import PwCheck from "./pages/PwCheck";
-import SignupSucess from "./pages/SignUpSuccess";
-import PwChangeSuccess from "./pages/PwChangeSuccess";
-import PwChange from "./pages/PwChange";
+import EmailSend from "./pages/signup/EmailSend";
+import PwSend from "./pages/signup/PwSend";
+import PwCheck from "./pages/signup/PwCheck";
+import SignupSucess from "./pages/signup/SignUpSuccess";
+import PwChangeSuccess from "./pages/signup/PwChangeSuccess";
+import PwChange from "./pages/signup/PwChange";
 import Main from "./pages/Main";
-import Job from "./pages/Job";
-import JobDetail from "./pages/JobDetail";
-import JobCategory from "./pages/JobCategory";
+import Job from "./pages/job/Job";
+import JobDetail from "./pages/job/JobDetail";
+import JobCategory from "./pages/job/JobCategory";
 import AddSchedule from "./components/AddSchedule";
 import PostDetail from "./pages/PostDetail";
 import Mypage from "./pages/Mypage";
