+++ src/pages/Main.jsx
@@ -5,14 +5,8 @@ import { useNavigate } from "react-router-dom";
 
 // 컴포넌트
 import Nav from "../components/Nav";
-import WeekSchedule from "../components/WeekSchedule";
-import MonthSchedule from "./../components/month/MonthSchedule";
-import MonthList from "../components/MonthList";
-import DailyList from "../components/DailyList";
 
 // 이미지
-import zoomin from "../assets/img/icon/zoomin.svg";
-import zoomout from "../assets/img/icon/zoomout.svg";
 import locationGray from "../assets/img/icon/LocationGray.svg";
 import searchImg from "../assets/img/icon/search.svg";
 import guideImg from "../assets/img/guide.svg";
@@ -29,17 +23,11 @@ import img8 from "../assets/img/sticker/sticker8.png";
 // 리덕스
 import { searchMySchedule } from "../redux/modules/search";
 import { useDispatch, useSelector } from "react-redux";
-import { zoomDate } from "../redux/modules/date";
-import WeekList from "../components/WeekList";
+import Calendar from "../components/Calendar";
 
 function Main() {
   const navigate = useNavigate();
   const dispatch = useDispatch();
-
-  const [weekMonth, setWeekMonth] = useState(true);
-  const [zoomInOut, setZoomInOut] = useState(true);
-  const [isActive, setIsActive] = useState(false);
-
   const navData = true;
 
   //일정등록 이동
@@ -90,11 +78,6 @@ function Main() {
     return dayOfWeek;
   }
 
-  // 월간달력 줌인 줌아웃
-  useEffect(() => {
-    dispatch(zoomDate(zoomInOut));
-  }, [zoomInOut]);
-
   const [showModal, setShowModal] = useState(false);
   const HAS_VISITED_BEFORE = localStorage.getItem("hasVisitedBefore");
 
@@ -140,55 +123,7 @@ function Main() {
         <AddButtoon onClick={MoveBtn}>+</AddButtoon>
       </FixBox>
       {search === "" ? (
-        <ContentWrap>
-          <ToggleBtn>
-            {weekMonth ? (
-              <>
-                {zoomInOut ? (
-                  <button
-                    onClick={() => {
-                      setZoomInOut(!zoomInOut);
-                    }}
-                  >
-                    <img src={zoomin} alt="월별달력확대" />
-                  </button>
-                ) : (
-                  <button
-                    onClick={() => {
-                      setZoomInOut(!zoomInOut);
-                    }}
-                  >
-                    <img src={zoomout} alt="월별달력축소" />
-                  </button>
-                )}
-              </>
-            ) : (
-              <div></div>
-            )}
-            {/* <WeekMonth
-              weekMonth={weekMonth}
-              onClick={() => {
-                setWeekMonth(!weekMonth);
-              }}
-            >
-              {weekMonth ? (
-                <Circle weekMonth={weekMonth}>M</Circle>
-              ) : (
-                <Circle weekMonth={weekMonth}>W</Circle>
-              )}
-            </WeekMonth> */}
-          </ToggleBtn>
-          {weekMonth ? (
-            <MonthSchedule isActive={isActive} setIsActive={setIsActive} />
-          ) : (
-            <WeekSchedule
-              weekMonth={weekMonth}
-              isActive={isActive}
-              setIsActive={setIsActive}
-            />
-          )}
-          {isActive ? <DailyList /> : weekMonth ? <MonthList /> : <WeekList />}
-        </ContentWrap>
+        <Calendar />
       ) : (
         <SearchWrapper>
           <UpBar>
@@ -308,45 +243,6 @@ const GuideBg = styled.div`
 
 const MainWrap = styled.div`
   position: relative;
-`;
-const ContentWrap = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: 18px 12px;
-  background-color: var(--blue1);
-  overflow: hidden;
-  overflow-y: scroll;
-  height: calc(100vh - 200px);
-`;
-const ToggleBtn = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-const WeekMonth = styled.button`
-  height: 27px;
-  width: 52px;
-  border-radius: 13px;
-  padding: 0 5px;
-  background-color: var(--blue2);
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  box-shadow: 0px 0px 8px 0px #3284ff45 inset;
-  box-shadow: 0px -1px 2px 0px #ffffff75;
-  justify-content: ${(props) => (props.weekMonth ? "flex-end" : "flex-start")};
-`;
-const Circle = styled.span`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-weight: 500;
-  font-size: 8px;
-  width: 20px;
-  height: 20px;
-  border-radius: 100%;
-  color: ${(props) => (props.weekMonth ? "var(--blue4)" : "#fff")};
-  background-color: ${(props) => (props.weekMonth ? "#fff" : "var(--blue4)")};
 `;
 const FixBox = styled.div`
   width: calc(100% - 24px);
