+++ src/pages/job/JobCategory.jsx
@@ -2,46 +2,50 @@ import React, { useEffect } from "react";
 import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { loadJobList, loadCategoryList, selectCategory } from "../redux/modules/job";
+import {
+  loadJobList,
+  loadCategoryList,
+  selectCategory,
+} from "../../redux/modules/job";
 
-import backBtn from "../assets/img/icon/Back.svg";
-import open from "../assets/img/btn/open.png";
-import close from "../assets/img/btn/close.png";
+import backBtn from "../../assets/img/icon/Back.svg";
+import open from "../../assets/img/btn/open.png";
+import close from "../../assets/img/btn/close.png";
 
 // 상세 지역 import
-import { Seoul } from "../shared/region";
-import { Gyeonggi } from "../shared/region";
-import { Incheon } from "../shared/region";
-import { Daejeon } from "../shared/region";
-import { Sejong } from "../shared/region";
-import { Chungnam } from "../shared/region";
-import { Chungbuk } from "../shared/region";
-import { Gwangju } from "../shared/region";
-import { Jeonnam } from "../shared/region";
-import { Jeonbuk } from "../shared/region";
-import { Daegu } from "../shared/region";
-import { Gyeongbuk } from "../shared/region";
-import { Busan } from "../shared/region";
-import { Ulsan } from "../shared/region";
-import { Gyeongnam } from "../shared/region";
-import { Gangwon } from "../shared/region";
-import { Jeju } from "../shared/region";
+import { Seoul } from "../../shared/region";
+import { Gyeonggi } from "../../shared/region";
+import { Incheon } from "../../shared/region";
+import { Daejeon } from "../../shared/region";
+import { Sejong } from "../../shared/region";
+import { Chungnam } from "../../shared/region";
+import { Chungbuk } from "../../shared/region";
+import { Gwangju } from "../../shared/region";
+import { Jeonnam } from "../../shared/region";
+import { Jeonbuk } from "../../shared/region";
+import { Daegu } from "../../shared/region";
+import { Gyeongbuk } from "../../shared/region";
+import { Busan } from "../../shared/region";
+import { Ulsan } from "../../shared/region";
+import { Gyeongnam } from "../../shared/region";
+import { Gangwon } from "../../shared/region";
+import { Jeju } from "../../shared/region";
 
 // 상세 직군 import
-import { 경영·사무 } from "../shared/jobData";
-import { 마케팅·광고·홍보 } from "../shared/jobData";
-import { IT·인터넷 } from "../shared/jobData";
-import { 디자인 } from "../shared/jobData";
-import { 무역·유통 } from "../shared/jobData";
-import { 영업·고객상담 } from "../shared/jobData";
-import { 서비스 } from "../shared/jobData";
-import { 연구개발·설계 } from "../shared/jobData";
-import { 생산·제조 } from "../shared/jobData";
-import { 교육 } from "../shared/jobData";
-import { 건설 } from "../shared/jobData";
-import { 의료 } from "../shared/jobData";
-import { 미디어 } from "../shared/jobData";
-import { 전문·특수직 } from "../shared/jobData";
+import { 경영·사무 } from "../../shared/jobData";
+import { 마케팅·광고·홍보 } from "../../shared/jobData";
+import { IT·인터넷 } from "../../shared/jobData";
+import { 디자인 } from "../../shared/jobData";
+import { 무역·유통 } from "../../shared/jobData";
+import { 영업·고객상담 } from "../../shared/jobData";
+import { 서비스 } from "../../shared/jobData";
+import { 연구개발·설계 } from "../../shared/jobData";
+import { 생산·제조 } from "../../shared/jobData";
+import { 교육 } from "../../shared/jobData";
+import { 건설 } from "../../shared/jobData";
+import { 의료 } from "../../shared/jobData";
+import { 미디어 } from "../../shared/jobData";
+import { 전문·특수직 } from "../../shared/jobData";
 
 const JobCategory = () => {
   const navigate = useNavigate();
@@ -92,7 +96,13 @@ const JobCategory = () => {
 
   const careers = ["신입", "경력", "신입·경력"];
 
-  const CompanyType = ["대기업", "중소/중견기업", "외국계기업", "공기업", "전체"];
+  const CompanyType = [
+    "대기업",
+    "중소/중견기업",
+    "외국계기업",
+    "공기업",
+    "전체",
+  ];
 
   const defaultCategory = useSelector((state) => state.job.category);
   const defaultCategor1y = useSelector((state) => state);
@@ -139,7 +149,11 @@ const JobCategory = () => {
       <CategoryTap>
         <TasksBtnWrap onClick={() => setTaskToggle(!taskToggle)}>
           <TasksBtnText>직무</TasksBtnText>
-          {taskToggle === false ? <TasksBtnImg src={close} /> : <TasksBtnImg src={open} />}
+          {taskToggle === false ? (
+            <TasksBtnImg src={close} />
+          ) : (
+            <TasksBtnImg src={open} />
+          )}
         </TasksBtnWrap>
         {taskToggle === false ? (
           <></>
@@ -174,7 +188,13 @@ const JobCategory = () => {
       ) : (
         <DetailJobTagsFlex>
           {categoryData.jobMain === "전체" ? (
-            <DetailJobTags categoryData={categoryData.jobSub} tasksData={"전체"} onClick={() => setCategoryData({ ...categoryData, jobSub: "전체" })}>
+            <DetailJobTags
+              categoryData={categoryData.jobSub}
+              tasksData={"전체"}
+              onClick={() =>
+                setCategoryData({ ...categoryData, jobSub: "전체" })
+              }
+            >
               {"전체"}
             </DetailJobTags>
           ) : categoryData.jobMain === "경영·사무" ? (
@@ -184,7 +204,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -197,7 +219,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -210,7 +234,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -223,7 +249,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -236,7 +264,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -249,7 +279,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -262,7 +294,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -275,7 +309,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -288,7 +324,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -301,7 +339,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobMain}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -314,7 +354,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobMain}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -327,7 +369,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -340,7 +384,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -353,7 +399,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.jobSub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, jobSub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, jobSub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailJobTags>
@@ -368,7 +416,11 @@ const JobCategory = () => {
       <CategoryTap>
         <TasksBtnWrap onClick={() => setworkAreaToggle(!workAreaToggle)}>
           <TasksBtnText>근무지역</TasksBtnText>
-          {workAreaToggle === false ? <TasksBtnImg src={close} /> : <TasksBtnImg src={open} />}
+          {workAreaToggle === false ? (
+            <TasksBtnImg src={close} />
+          ) : (
+            <TasksBtnImg src={open} />
+          )}
         </TasksBtnWrap>
         {workAreaToggle === false ? (
           <></>
@@ -403,7 +455,13 @@ const JobCategory = () => {
       ) : (
         <DetailWorkAreaTagsFlex>
           {categoryData.cityMain === "전체" ? (
-            <DetailWorkAreaTags categoryData={categoryData.citySub} tasksData={"전체"} onClick={() => setCategoryData({ ...categoryData, citySub: "전체" })}>
+            <DetailWorkAreaTags
+              categoryData={categoryData.citySub}
+              tasksData={"전체"}
+              onClick={() =>
+                setCategoryData({ ...categoryData, citySub: "전체" })
+              }
+            >
               {"전체"}
             </DetailWorkAreaTags>
           ) : categoryData.cityMain === "서울" ? (
@@ -413,7 +471,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -426,7 +486,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -439,7 +501,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -452,7 +516,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -465,7 +531,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -478,7 +546,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -491,7 +561,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -504,7 +576,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -517,7 +591,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -530,7 +606,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -543,7 +621,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -556,7 +636,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -569,7 +651,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -582,7 +666,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -595,7 +681,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -608,7 +696,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -621,7 +711,9 @@ const JobCategory = () => {
                   key={idx}
                   categoryData={categoryData.citySub}
                   tasksData={workAreaData}
-                  onClick={() => setCategoryData({ ...categoryData, citySub: workAreaData })}
+                  onClick={() =>
+                    setCategoryData({ ...categoryData, citySub: workAreaData })
+                  }
                 >
                   {workAreaData}
                 </DetailWorkAreaTags>
@@ -641,9 +733,19 @@ const JobCategory = () => {
             return (
               <CareersTags
                 key={idx}
-                categoryData={categoryData.career === "경력무관" ? "신입·경력" : categoryData.career}
+                categoryData={
+                  categoryData.career === "경력무관"
+                    ? "신입·경력"
+                    : categoryData.career
+                }
                 tasksData={careersData}
-                onClick={() => setCategoryData({ ...categoryData, career: careersData === "신입·경력" ? "경력무관" : careersData })}
+                onClick={() =>
+                  setCategoryData({
+                    ...categoryData,
+                    career:
+                      careersData === "신입·경력" ? "경력무관" : careersData,
+                  })
+                }
               >
                 {careersData}
               </CareersTags>
@@ -661,7 +763,12 @@ const JobCategory = () => {
                 key={idx}
                 categoryData={categoryData.companyType}
                 tasksData={CompanyTypeData}
-                onClick={() => window.alert("준비하고 있는 기능입니다!", setCategoryData({ ...categoryData, companyType: "대기업" }))}
+                onClick={() =>
+                  window.alert(
+                    "준비하고 있는 기능입니다!",
+                    setCategoryData({ ...categoryData, companyType: "대기업" })
+                  )
+                }
               >
                 {CompanyTypeData}
               </CompanyTypeTags>
@@ -748,7 +855,8 @@ const TasksBtnImg = styled.img`
 const JobTags = styled.div`
   padding: 4px 10px;
   display: inline-block;
-  background: ${(props) => (props.categoryData === props.tasksData ? "var(--blue4)" : "transparent")};
+  background: ${(props) =>
+    props.categoryData === props.tasksData ? "var(--blue4)" : "transparent"};
   border-radius: 6px;
   :nth-child(1) {
     margin-right: calc(100% - 50px);
@@ -758,7 +866,8 @@ const JobTags = styled.div`
   line-height: 20px;
   font-weight: 600;
   font-size: 14px;
-  color: ${(props) => (props.categoryData === props.tasksData ? "white" : "var(--blue4)")};
+  color: ${(props) =>
+    props.categoryData === props.tasksData ? "white" : "var(--blue4)"};
   cursor: pointer;
 `;
 const JobTagsFlex = styled.div`
@@ -773,12 +882,14 @@ const DetailJobTags = styled.div`
   padding: 6px 10px;
   margin-bottom: 16px;
   margin-right: 20px;
-  background: ${(props) => (props.categoryData === props.tasksData ? "var(--blue4)" : "transparent")};
+  background: ${(props) =>
+    props.categoryData === props.tasksData ? "var(--blue4)" : "transparent"};
   border-radius: 8px;
   line-height: 20px;
   font-weight: 700;
   font-size: 12px;
-  color: ${(props) => (props.categoryData === props.tasksData ? "white" : "#74A0E3")};
+  color: ${(props) =>
+    props.categoryData === props.tasksData ? "white" : "#74A0E3"};
   border: 1px solid var(--blue2);
   box-sizing: border-box;
   cursor: pointer;
@@ -786,25 +897,29 @@ const DetailJobTags = styled.div`
 
 const WorkAreaTags = styled.div`
   padding: 8px 10px;
-  background: ${(props) => (props.categoryData === props.tasksData ? "var(--blue4)" : "transparent")};
+  background: ${(props) =>
+    props.categoryData === props.tasksData ? "var(--blue4)" : "transparent"};
   border-radius: 6px;
   font-size: 14px;
   line-height: 20px;
   font-weight: 600;
   font-size: 14px;
-  color: ${(props) => (props.categoryData === props.tasksData ? "white" : "var(--blue4)")};
+  color: ${(props) =>
+    props.categoryData === props.tasksData ? "white" : "var(--blue4)"};
   cursor: pointer;
 `;
 
 const DetailWorkAreaTags = styled.div`
   padding: 6px 10px;
   margin: 4px;
-  background: ${(props) => (props.categoryData === props.tasksData ? "var(--blue4)" : "transparent")};
+  background: ${(props) =>
+    props.categoryData === props.tasksData ? "var(--blue4)" : "transparent"};
   border-radius: 8px;
   line-height: 20px;
   font-weight: 700;
   font-size: 12px;
-  color: ${(props) => (props.categoryData === props.tasksData ? "white" : "#74A0E3")};
+  color: ${(props) =>
+    props.categoryData === props.tasksData ? "white" : "#74A0E3"};
   border: 1px solid var(--blue2);
   box-sizing: border-box;
   cursor: pointer;
@@ -830,12 +945,14 @@ const CareersTags = styled.div`
   padding: 6px 10px;
   margin: 4px;
   margin-bottom: 8px;
-  background: ${(props) => (props.categoryData === props.tasksData ? "var(--blue4)" : "transparent")};
+  background: ${(props) =>
+    props.categoryData === props.tasksData ? "var(--blue4)" : "transparent"};
   border-radius: 8px;
   line-height: 20px;
   font-weight: 700;
   font-size: 12px;
-  color: ${(props) => (props.categoryData === props.tasksData ? "white" : "var(--blue3)")};
+  color: ${(props) =>
+    props.categoryData === props.tasksData ? "white" : "var(--blue3)"};
   border: 1px solid var(--blue2);
   box-sizing: border-box;
   cursor: pointer;
@@ -846,12 +963,14 @@ const CompanyTypeTags = styled.div`
   padding: 4px 10px;
   margin: 4px;
   margin-bottom: 8px;
-  background: ${(props) => (props.categoryData === props.tasksData ? "var(--blue4)" : "transparent")};
+  background: ${(props) =>
+    props.categoryData === props.tasksData ? "var(--blue4)" : "transparent"};
   border-radius: 8px;
   line-height: 20px;
   font-weight: 700;
   font-size: 12px;
-  color: ${(props) => (props.categoryData === props.tasksData ? "white" : "var(--blue3)")};
+  color: ${(props) =>
+    props.categoryData === props.tasksData ? "white" : "var(--blue3)"};
   border: 1px solid var(--blue2);
   box-sizing: border-box;
   cursor: pointer;
