+++ src/pages/signup/SignUpSuccess.jsx
@@ -1,27 +1,40 @@
 import React from "react";
 import styled from "styled-components";
-import { Link } from "react-router-dom";
-import passwordChangeImg from "../assets/img/illust/passwordchange.png";
-const PwChangeSuccess = () => {
+import { useSelector } from "react-redux";
+import { useNavigate } from "react-router-dom";
+import signUpImg from "../../assets/img/illust/signupsuccess.png";
+
+//회원이름
+const SignupSucess = () => {
+  const navigate = useNavigate();
+  const userInfo = useSelector((state) => state.user.user);
+
   return (
     <EmailWrap>
       <Header>
-        <Banner src={passwordChangeImg} alt="배너" />
+        <Banner src={signUpImg} alt="배너" />
         <TitleText>
-          <Title>비밀번호가 변경되었어요!</Title>
-          <SubTitle>다시 한번 힘차게 로그인하러 가볼까요?</SubTitle>
+          <Title>회원가입이 완료되었어요!</Title>
+          <SubTitle>
+            {userInfo.userName} 유저님의 취준 여정에 <br />
+            굿잡캘린더가 함께 할게요!
+          </SubTitle>
         </TitleText>
       </Header>
       <Main>
-        <SignUpBtn>
-          <Link to="/login">로그인하기</Link>
+        <SignUpBtn
+          onClick={() => {
+            navigate("/login");
+          }}
+        >
+          로그인하기
         </SignUpBtn>
       </Main>
     </EmailWrap>
   );
 };
 
-export default PwChangeSuccess;
+export default SignupSucess;
 const EmailWrap = styled.div`
   display: flex;
   flex-direction: column;
@@ -46,7 +59,9 @@ const Header = styled.header`
   position: relative;
   width: 100%;
   padding-bottom: 50px;
-  margin-bottom: 73px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 `;
 const TitleText = styled.div`
   position: absolute;
@@ -55,6 +70,7 @@ const TitleText = styled.div`
   top: 70%;
   width: 100%;
   text-align: center;
+  z-index: 999;
 `;
 const Title = styled.h1`
   font-weight: 700;
@@ -67,8 +83,9 @@ const SubTitle = styled.p`
   margin-top: 16px;
 `;
 const Banner = styled.img`
-  width: 100%;
+  width: 60%;
   border-radius: 26px;
+  margin-bottom: 16px;
 `;
 const Main = styled.main`
   display: flex;
@@ -84,8 +101,7 @@ const SignUpBtn = styled.button`
   display: flex;
   justify-content: center;
   align-items: center;
-  margin-bottom: 78px;
   font-weight: 400;
   color: #fff !important;
-  margin-top: 72px;
+  margin-top: 68px;
 `;
